pyproject.toml
[tool.ruff]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "W",  # pycodestyle warnings
  "UP", # pyupgrade
]
ignore = [
  "E203",  # whitespace before ':' (black compat)
  "E501",  # line length (handled separately)
]
line-length = 120
target-version = "py310"
exclude = [
  ".git",
  "__pycache__",
  "dist",
  "build",
  "venv",
  ".venv",
  "ui",
  "languages",
  ".github",
]

[tool.ruff.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["utils", "API_SDK", "engine_sdk", "bcasl", "acasl"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = false
warn_return_any = false
no_implicit_optional = false
pretty = true
show_error_codes = true
exclude = "(venv|.venv|dist|build|ui|languages|.github)"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["utils", "API_SDK", "engine_sdk", "bcasl", "acasl"]

[tool.coverage.report]
omit = [
  "*/tests/*",
  "*/venv/*",
  "*/.venv/*",
  "*/ui/*",
  "*/languages/*",
  "*/.github/*",
]
show_missing = true
skip_covered = true
precision = 2

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "ui", "languages", ".github"]
skips = ["B101", "B601"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ui
  | languages
  | \.github
)/
'''

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pycompiler-ark"
version = "3.2.3"
description = "PyCompiler ARK++ - Advanced Python compilation toolkit with plugin architecture"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "PyCompiler ARK++ Team", email = "team@pycompiler-ark.org"}
]
maintainers = [
    {name = "PyCompiler ARK++ Maintainers", email = "maintainers@pycompiler-ark.org"}
]
keywords = ["python", "compiler", "packaging", "build-tools", "pyinstaller", "nuitka"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Compilers",
    "Topic :: System :: Software Distribution",
]
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.7.2,<6.8.0",
    "shiboken6>=6.7.2,<6.8.0",
    "psutil>=5.9.0,<6.0.0",
    "PyYAML>=6.0.0,<7.0.0",
    "tomli>=2.0.1,<3.0.0; python_version < '3.11'",
    "jsonschema>=4.18.0,<5.0.0",
    "Pillow>=9.0.0,<11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-qt>=4.2.0,<5.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.12.0,<24.0.0",
    "ruff>=0.1.9,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "bandit>=1.7.5,<2.0.0",
    "pip-audit>=2.6.0,<3.0.0",
    "safety>=2.3.0,<3.0.0",
    "pre-commit>=3.0.0,<4.0.0",
]
build = [
    "build>=1.0.0,<2.0.0",
    "wheel>=0.42.0,<1.0.0",
    "PyInstaller>=6.3.0,<7.0.0",
    "Nuitka>=1.9.0,<2.0.0",
]
security = [
    "cyclonedx-py>=4.0.0,<5.0.0",
    "bandit>=1.7.5,<2.0.0",
    "pip-audit>=2.6.0,<3.0.0",
    "safety>=2.3.0,<3.0.0",
]
docs = [
    "sphinx>=7.0.0,<8.0.0",
    "sphinx-rtd-theme>=1.3.0,<2.0.0",
    "myst-parser>=2.0.0,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/pycompiler-ark"
Documentation = "https://pycompiler-ark.readthedocs.io"
Repository = "https://github.com/your-org/pycompiler-ark.git"
Issues = "https://github.com/your-org/pycompiler-ark/issues"
Changelog = "https://github.com/your-org/pycompiler-ark/blob/main/CHANGELOG.md"
Security = "https://github.com/your-org/pycompiler-ark/security"

[project.scripts]
pycompiler-ark = "main:main"

[project.gui-scripts]
pycompiler-ark-gui = "main:main"